canonical/pyproject.toml
[project]
name = "canonical"
version = "1.0.0"
description = "Canonical schema library for RM Brain - shared entity schemas, event schemas, and semantic constraints"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Package is in src/canonical/ directory (standard Python package structure)
# Hatchling automatically maps src/canonical/ to canonical/ in the wheel
# Data directories (entities, events, semantics) are now in src/canonical/
# so they'll be included automatically as part of the package
packages = ["src/canonical"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
